+++ app/jsx/components/groups/show_map.jsx
