+++ app/jsx/components/restaurants/restaurants_list.js
@@ -1,7 +1,7 @@
 let translate = require('counterpart');
 
 import axios from 'axios';
-import * as constant from  '../constant';
+import * as constant from  '../../constant';
 import Restaurant from './restaurant';
 
 export default class RestaurantsList extends React.Component {
