+++ app/jsx/components/static_pages/footer.jsx
