+++ app/jsx/components/static_pages/jumbotron.jsx
