+++ app/jsx/components/static_pages/our_story.jsx
