+++ app/jsx/components/static_pages/quick_intro.jsx
