+++ app/jsx/components/static_pages/static_pages.jsx
