+++ app/jsx/components/users/login.jsx
@@ -1,33 +1,37 @@
 let translate = require('counterpart');
 import axios from 'axios';
-import * as constant from  '../constant';
+import * as constant from  '../../constant';
 import {Route, Link} from 'react-router';
 
-export default class Signup extends React.Component {
+export default class Login extends React.Component {
   constructor(props) {
     super(props);
-    this.state = {email: '', password: '', password_confirmation: ''};
+    this.state = {email: '', password: '', remember_me: true};
+  }
+
+  componentWillMount() {
+    if (localStorage.feastival_user != null) {
+      window.location = constant.BASE_URL;
+    }
   }
 
   handleSubmit(event) {
     event.preventDefault();
     let formData = new FormData();
-    formData.append('user[email]', this.state.email);
-    formData.append('user[password]', this.state.password);
-    formData.append('user[password_confirmation]', this.state.password);
+    formData.append('sign_in[email]', this.state.email);
+    formData.append('sign_in[password]', this.state.password);
     let email = this.state.email;
-
-    axios.post(constant.API_SIGN_UP_URL, formData)
-      .then(response =>  {
+    axios.post(constant.API_SIGN_IN_URL, formData)
+      .then((response) =>  {
         let feastival_user = {
           email: email,
-          user_id: response.data.user.id,
-          USER_TOKEN: response.data.user.authentication_token
+          user_id: response.data.user_session.id,
+          USER_TOKEN: response.data.user_session.user_token
         }
         localStorage.setItem('feastival_user', JSON.stringify(feastival_user));
-        window.location = constant.CURRENT_USER_INFO_URL;
+        window.location = constant.BASE_URL;
       })
-      .catch(error => {
+      .catch(function (error) {
         alert(error);
       });
   }
@@ -42,9 +46,9 @@ export default class Signup extends React.Component {
     });
   }
 
-  signInClick() {
+  signUpClick() {
     this.preventDefault;
-    window.location = constant.SIGN_IN_URL
+    window.location = constant.SIGN_UP_URL
   }
 
   render() {
@@ -55,12 +59,12 @@ export default class Signup extends React.Component {
             <div className='login-card'>
               <form onSubmit={this.handleSubmit.bind(this)} method='post'>
                 <div className='pmd-card-title card-header-border text-center'>
-                  <h3><strong>{translate('app.signup.register')}</strong></h3>
+                  <h3><strong>{translate('app.login.sign_in')}</strong></h3>
                 </div>
 
                 <div className='pmd-card-body'>
                   <div className='alert alert-success report' role='alert'>
-                    {translate('app.signup.error')}</div>
+                    {translate('app.login.error')}</div>
                   <div className='form-group pmd-textfield pmd-textfield-floating-label'>
                     <div className='input-group'>
                       <div className='input-group-addon'>
@@ -70,7 +74,7 @@ export default class Signup extends React.Component {
                         value={this.state.email}
                         name='email'
                         onChange={this.handleInputChange.bind(this)}
-                        placeholder={translate('app.signup.email')}
+                        placeholder={translate('app.login.email')}
                       />
                     </div>
                   </div>
@@ -84,39 +88,21 @@ export default class Signup extends React.Component {
                         value={this.state.password}
                         name='password'
                         onChange={this.handleInputChange.bind(this)}
-                        placeholder={translate('app.signup.password')}
+                        placeholder={translate('app.login.password')}
                       />
                     </div>
                   </div>
-
-                  <div className='form-group pmd-textfield pmd-textfield-floating-label'>
-                    <div className='input-group'>
-                      <div className='input-group-addon'>
-                        <i className='material-icons md-dark pmd-sm'>lock_outline</i>
-                      </div>
-                      <input type='password' className='form-control'
-                        value={this.state.password_confirmation}
-                        name='password_confirmation'
-                        onChange={this.handleInputChange.bind(this)}
-                        placeholder={translate('app.signup.password_confirm')}
-                      />
-                    </div>
-                  </div>
-
                 </div>
                 <div className='pmd-card-footer card-footer-no-border card-footer-p16 text-center'>
-
                   <button className='btn btn-primary btn-block' type='submit'>
-                    {translate('app.signup.submit')}</button>
-
-                  <p className='redirection-link'>
-                    {translate('app.signup.got_account')}
-                    <Link to='/login'>{translate('app.signup.login')}</Link>
+                    {translate('app.login.login_view')}
+                  </button>
+                  <p className='redirection-link'>{translate('app.login.not_user')}
+                    <Link to={'/signup'}>{translate('app.login.sign_up')}</Link>
                   </p>
                 </div>
               </form>
             </div>
-
           </div>
         </div>
       </section>
