+++ app/jsx/components/users/user_info.jsx
@@ -1,6 +1,6 @@
 let translate = require('counterpart');
 import update from 'react-addons-update';
-import * as constant from  '../constant'
+import * as constant from  '../../constant'
 import axios from 'axios';
 
 export default class UserInfo extends React.Component {
